並べ替え

+++ utils/metadata.py
@@ -2,13 +2,22 @@ import bpy
 import os
 import json
 from typing import Union
-from ..constants import METADATA_FILENAME_PREFIX, STORAGE_MACROS_DIR
+from ..state import state
+from ..constants import (
+    METADATA_FILENAME_PREFIX,
+    REGION_GLOBAL,
+    REGION_LOCAL,
+    STORAGE_MACROS_DIR,
+)
 
-def to_dict(data:str):
+
+def to_dict(data: str):
     return json.loads(data)
 
-def to_json(data:dict):
+
+def to_json(data: dict):
     return json.dumps(data)
+
 
 def get_local_metadata(name: str):
     _name = METADATA_FILENAME_PREFIX + name
@@ -63,14 +72,6 @@ def rename_local_metadata(old_name: str, new_name: str):
     macro.name = METADATA_FILENAME_PREFIX + new_name
 
 
-
-
-
-
-
-
-
-
 def write_to_global_metadata(name: str, data: dict):
     path = f"{STORAGE_MACROS_DIR}/{METADATA_FILENAME_PREFIX}{name}.py"
 
@@ -80,20 +81,29 @@ def write_to_global_metadata(name: str, data: dict):
     with open(path, "w") as f:
         f.write(to_json(data))
 
+    state["global_metadata_caches"][name] = data
 
-def read_from_global_metadata(name: str) -> Union[dict, None]:
+
+def read_from_global_metadata(name: str, update=False) -> Union[dict, None]:
     path = f"{STORAGE_MACROS_DIR}/{METADATA_FILENAME_PREFIX}{name}.py"
+
+    if not update and name in state["global_metadata_caches"]:
+        return state["global_metadata_caches"][name]
 
     if not os.path.exists(path):
         return None
 
     with open(path) as f:
-        return to_dict(f.read())
+        data = to_dict(f.read())
+        state["global_metadata_caches"][name] = data
+
+    return data
 
 
 def remove_global_metadata(name: str):
     path = f"{STORAGE_MACROS_DIR}/{METADATA_FILENAME_PREFIX}{name}.py"
-    os.remove(path)
+    if os.path.exists(path):
+        os.remove(path)
 
 
 def rename_global_metadata(old_name: str, new_name: str):
@@ -103,7 +113,7 @@ def rename_global_metadata(old_name: str, new_name: str):
 
 
 def move_metadata_to_global(local_metadata_name: str):
-    global_metadata_data = read_from_global_metadata(local_metadata_name)
+    global_metadata_data = read_from_global_metadata(local_metadata_name, True)
     if global_metadata_data is not None:
         raise ValueError("Already exists!")
 
@@ -111,6 +121,7 @@ def move_metadata_to_global(local_metadata_name: str):
     if data is not None:
         write_to_global_metadata(local_metadata_name, data)
         remove_local_metadata(local_metadata_name)
+
 
 def move_metadata_to_local(global_metadata_name: str):
     local_macro_data = read_from_local_metadata(global_metadata_name)
@@ -121,4 +132,20 @@ def move_metadata_to_local(global_metadata_name: str):
     if data is not None:
         write_to_local_metadata(global_metadata_name, data)
         remove_global_metadata(global_metadata_name)
+
+
+def read_metadata(name: str, region: str):
+    if region == REGION_LOCAL:
+        return read_from_local_metadata(name)
+
+    if region == REGION_GLOBAL:
+        return read_from_global_metadata(name)
+
+
+def write_metadata(name: str, data: dict, region: str):
+    if region == REGION_LOCAL:
+        write_to_local_metadata(name, data)
+
+    if region == REGION_GLOBAL:
+        write_to_global_metadata(name, data)
